+++ src/config.lua
@@ -8,8 +8,12 @@ local config = {
     oceanus = true;
     fields = true;
     tartarus = true;
+
     ephyra = true;
     ephyra_sideroom = true;
+
+    asphodel = true;
+    elysium = true;
   };
   heracles = {
     enabled = true;
@@ -19,17 +23,23 @@ local config = {
     oceanus = true;
     fields = true;
     tartarus = true;
+
+    tartarus_nightmare = true;
+    asphodel = true;
+    elysium = true;
   };
   artemis = {
     enabled = true;
     weight = 1;
 
-    erebus = true;
-    oceanus = true;
     fields = true;
     tartarus = true;
+
     thessaly = true;
     olympus = true;
+
+    tartarus_nightmare = true;
+    elysium = true;
   };
   nemesis = {
     enabled = true;
@@ -38,6 +48,10 @@ local config = {
     ephyra = true;
     thessaly = true;
     olympus = true;
+
+    tartarus_nightmare = true;
+    asphodel = true;
+    elysium = true;
   };
   thanatos = {
     enabled = true;
@@ -48,7 +62,9 @@ local config = {
     tartarus = true;
     ephyra = true;
     olympus = true;
-  }
+  };
+
+  zags_journey_integration = true;
 }
 
 local configDesc = {
@@ -61,8 +77,12 @@ local configDesc = {
     oceanus = "Allow Icarus in Oceanus";
     fields = "Allow Icarus in Fields";
     tartarus = "Allow Icarus in Tartarus";
+
     ephyra = "Allow Icarus in Ephyra";
     ephyra_sideroom = "Allow Icarus in Ephyra side rooms";
+
+    asphodel = "Allow Icarus in Asphodel";
+    elysium = "Allow Icarus in Elysium";
   };
   heracles = {
     enabled = "Enable/disable Extended Heracles Encounters";
@@ -72,17 +92,23 @@ local configDesc = {
     oceanus = "Allow Heracles in Oceanus";
     fields = "Allow Heracles in Fields";
     tartarus = "Allow Heracles in Tartarus";
+
+    tartarus_nightmare = "Allow Heracles in Tartarus(Nightmare)";
+    asphodel = "Allow Heracles in Asphodel";
+    elysium = "Allow Heracles in Elysium";
   };
   artemis = {
     enabled = "Enable/disable Extended Artemis Encounters";
     weight = "Weight range: 1-20";
 
-    erebus = "Allow Artemis in Erebus";
-    oceanus = "Allow Artemis in Oceanus";
     fields = "Allow Artemis in Fields";
     tartarus = "Allow Artemis in Tartarus";
+
     thessaly = "Allow Artemis in Thessaly";
     olympus = "Allow Artemis in Olympus";
+
+    tartarus_nightmare = "Allow Artemis in Tartarus(Nightmare)";
+    elysium = "Allow Artemis in Elysium";
   };
   nemesis = {
     enabled = "Enable/disable Extended Nemesis Encounters";
@@ -91,6 +117,10 @@ local configDesc = {
     ephyra = "Allow Nemesis in Ephyra";
     thessaly = "Allow Nemesis in Thessaly";
     olympus = "Allow Nemesis in Olympus";
+
+    tartarus_nightmare = "Allow Nemesis in Tartarus(Nightmare)";
+    asphodel = "Allow Nemesis in Asphodel";
+    elysium = "Allow Nemesis in Elysium";
   };
   thanatos = {
     enabled = "Enable/disable Extended Thanatos Encounters";
@@ -99,9 +129,12 @@ local configDesc = {
     erebus = "Allow Thanatos in Erebus";
     oceanus = "Allow Thanatos in Oceanus";
     tartarus = "Allow Thanatos in Tartarus";
+
     ephyra = "Allow Thanatos in Ephyra";
     olympus = "Allow Thanatos in Olympus";
-  }
+  };
+
+  zags_journey_integration = "Enable/Disable Zagreus' Journey integration";
 }
 
 return config, configDesc